# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Österreichische EEG (Energiegemeinschaft) Abrechnungsplattform — REST API",
        "title": "eegabrechnung API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "eegabrechnung Support",
            "email": "support@eeg.at"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8101",
    "basePath": "/api/v1",
    "paths": {
...
        "/eegs/{eegID}/tariffs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gibt alle Tarifpläne (TariffSchedules) einer EEG zurück.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifpläne auflisten",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste der Tarifpläne",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.TariffSchedule"
                            }
                        }
                    },
                    "400": {
                        "description": "Ungültige EEG ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Legt einen neuen Tarifplan für eine EEG an. Granularität: annual, monthly, daily, quarter_hour (Standard: monthly).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan anlegen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tarifplan-Daten",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Angelegter Tarifplan",
                        "schema": {
                            "$ref": "#/definitions/domain.TariffSchedule"
                        }
                    },
                    "400": {
                        "description": "Ungültige EEG ID oder fehlender Name",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/tariffs/{scheduleID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gibt einen einzelnen Tarifplan inklusive aller Tarifeinträge zurück.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan abrufen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarifplan mit Einträgen",
                        "schema": {
                            "$ref": "#/definitions/domain.TariffSchedule"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "404": {
                        "description": "Tarifplan nicht gefunden",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aktualisiert Name und Granularität eines bestehenden Tarifplans.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan aktualisieren",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Aktualisierte Tarifplan-Daten",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Aktualisierter Tarifplan",
                        "schema": {
                            "$ref": "#/definitions/domain.TariffSchedule"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID oder Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Löscht einen Tarifplan und alle zugehörigen Tarifeinträge endgültig.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan löschen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Bestätigung",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/tariffs/{scheduleID}/activate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Aktiviert einen Tarifplan für die EEG. Pro EEG kann nur ein Tarifplan gleichzeitig aktiv sein. Gibt 409 zurück, wenn bereits ein anderer aktiv ist.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan aktivieren",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "is_active: true",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "409": {
                        "description": "Ein anderer Tarifplan ist bereits aktiv",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deaktiviert den aktiven Tarifplan, ohne ihn zu löschen.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifplan deaktivieren",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "is_active: false",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/eegs/{eegID}/tariffs/{scheduleID}/entries": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ersetzt alle Tarifeinträge eines Tarifplans vollständig durch die übergebene Liste. Jeder Eintrag enthält einen Gültigkeitszeitraum sowie Energie- und Einspeisepreis in ct/kWh.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tarifpläne"
                ],
                "summary": "Tarifeinträge setzen",
                "parameters": [
                    {
                        "type": "string",
                        "description": "EEG ID (UUID)",
                        "name": "eegID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tarifplan-ID (UUID)",
                        "name": "scheduleID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Liste der Tarifeinträge",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anzahl gespeicherter Einträge",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "400": {
                        "description": "Ungültige ID, Datumsformat oder Request",
                        "schema": {
                            "type": "object"
                        }
                    },
                    "500": {
                        "description": "Interner Fehler",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.TariffEntry": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "energy_price": {
                    "description": "ct/kWh",
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "producer_price": {
                    "description": "ct/kWh",
                    "type": "number"
                },
                "schedule_id": {
                    "type": "string"
                },
                "valid_from": {
                    "type": "string"
                },
                "valid_until": {
                    "type": "string"
                }
            }
        },
        "domain.TariffSchedule": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "eeg_id": {
                    "type": "string"
                },
                "entries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.TariffEntry"
                    }
                },
                "entry_count": {
                    "type": "integer"
                },
                "granularity": {
                    "description": "annual | monthly | daily | quarter_hour",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "is_active": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "security": [
        {
            "BearerAuth": []
        }
    ]
}